pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-clip"
version = "0.1.0"
description = "Mini-CLIP research codebase"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "torchvision",
  "transformers",
  "pyyaml",
  "tqdm",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
